backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
email-validator==2.1.1
sqlalchemy>=2.0.35
psycopg[binary]>=3.2.0
alembic==1.13.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjwt==2.8.0

